src/requirements.txt
matplotlib~=3.6.2
fpdf~=1.7.2
Pillow~=9.3.0
pandas~=1.5.2
requests~=2.28.1
bs4~=0.0.1
beautifulsoup4~=4.11.1
dataframe_image~=0.1.3
fpdf2